refactor: DragAndDrop 로직 setup함수 분리
- Bookshelf의 DragAndDrop 로직을 setup 함수로 분리
