+++ src/newtab/components/composition/setupBookshelfAction.ts
@@ -4,7 +4,6 @@ import {
   SET_TOOLTIP_POSITION,
   SET_TOOLTIP_SHOW,
   SET_TOOLTIP_TEXT,
-  SET_TOOLTIP_ON,
 } from "@/newtab/store/modules/tooltip";
 import { Item } from "@/shared/types/store";
 import { Ref, SetupContext } from "vue";
@@ -23,8 +22,6 @@ export interface SetupBookshelf {
   openUrl: (url: string) => void;
   openContextMenu: (event: PointerEvent, target: ContextMenuTarget) => void;
   onClickFolder: (item: Item) => void;
-  offTooltip: () => void;
-  onTooltip: () => void;
 }
 
 export const setupBookshelfAction = (props: Props): SetupBookshelf => {
@@ -53,13 +50,7 @@ export const setupBookshelfAction = (props: Props): SetupBookshelf => {
   const closeTooltip = () => {
     store.commit(SET_TOOLTIP_SHOW, false);
   };
-  const onTooltip = () => {
-    store.commit(SET_TOOLTIP_ON, true);
-  };
 
-  const offTooltip = () => {
-    store.commit(SET_TOOLTIP_ON, false);
-  };
   const openUrl = (url: string) => {
     window.open(url, "_blank")?.focus();
     store.commit(SET_TOOLTIP_SHOW, false);
@@ -82,7 +73,5 @@ export const setupBookshelfAction = (props: Props): SetupBookshelf => {
     openUrl,
     openContextMenu,
     onClickFolder,
-    onTooltip,
-    offTooltip,
   };
 };
